+++ test/specs/openapi/test_openapi_checks.py
@@ -19,6 +19,6 @@ from schemathesis.specs.openapi.checks import (
 def test_wrong_schema_type(check):
     schema = BaseSchema({})
     # These checks should not be valid for some generic schema
-    case = Case(endpoint=APIOperation("", "", None, schema=schema))
+    case = Case(operation=APIOperation("", "", None, schema=schema))
     with pytest.raises(TypeError):
         check(requests.Response(), case)